DOPE-216: move test files to correct package
