+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/collection/NotWithinExpressionTest.kt
@@ -1,4 +1,4 @@
-package ch.ergon.dope.operators.collection
+package ch.ergon.dope.resolvable.expression.unaliased.type.collection
 
 import ch.ergon.dope.DopeParameters
 import ch.ergon.dope.DopeQuery
@@ -15,21 +15,21 @@ import ch.ergon.dope.helper.someStringArrayField
 import ch.ergon.dope.helper.someStringField
 import ch.ergon.dope.helper.someStringSelectRawClause
 import ch.ergon.dope.resolvable.expression.unaliased.type.asParameter
-import ch.ergon.dope.resolvable.expression.unaliased.type.collection.WithinExpression
-import ch.ergon.dope.resolvable.expression.unaliased.type.collection.withinArray
+import ch.ergon.dope.resolvable.expression.unaliased.type.collection.NotWithinExpression
+import ch.ergon.dope.resolvable.expression.unaliased.type.collection.notWithinArray
 import ch.ergon.dope.resolvable.expression.unaliased.type.toDopeType
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class WithinExpressionTest : ManagerDependentTest {
+class NotWithinExpressionTest : ManagerDependentTest {
     override lateinit var manager: DopeQueryManager
 
     @Test
-    fun `should support WITHIN expression`() {
+    fun `should support NOT WITHIN expression`() {
         val expected = DopeQuery(
-            queryString = "`numberField` WITHIN `numberArrayField`",
+            queryString = "`numberField` NOT WITHIN `numberArrayField`",
         )
-        val underTest = WithinExpression(someNumberField(), someNumberArrayField())
+        val underTest = NotWithinExpression(someNumberField(), someNumberArrayField())
 
         val actual = underTest.toDopeQuery(manager)
 
@@ -37,14 +37,14 @@ class WithinExpressionTest : ManagerDependentTest {
     }
 
     @Test
-    fun `should support WITHIN expression with named parameter as value`() {
+    fun `should support NOT WITHIN expression with named parameter as value`() {
         val parameterValue = 1
         val parameterName = "param"
         val expected = DopeQuery(
-            queryString = "\$$parameterName WITHIN `numberArrayField`",
+            queryString = "\$$parameterName NOT WITHIN `numberArrayField`",
             DopeParameters(namedParameters = mapOf(parameterName to parameterValue)),
         )
-        val underTest = WithinExpression(parameterValue.asParameter(parameterName), someNumberArrayField())
+        val underTest = NotWithinExpression(parameterValue.asParameter(parameterName), someNumberArrayField())
 
         val actual = underTest.toDopeQuery(manager)
 
@@ -52,13 +52,13 @@ class WithinExpressionTest : ManagerDependentTest {
     }
 
     @Test
-    fun `should support WITHIN expression with positional parameter as value`() {
+    fun `should support NOT WITHIN expression with positional parameter as value`() {
         val parameterValue = 1
         val expected = DopeQuery(
-            queryString = "$1 WITHIN `numberArrayField`",
+            queryString = "$1 NOT WITHIN `numberArrayField`",
             DopeParameters(positionalParameters = listOf(parameterValue)),
         )
-        val underTest = WithinExpression(parameterValue.asParameter(), someNumberArrayField())
+        val underTest = NotWithinExpression(parameterValue.asParameter(), someNumberArrayField())
 
         val actual = underTest.toDopeQuery(manager)
 
@@ -66,14 +66,14 @@ class WithinExpressionTest : ManagerDependentTest {
     }
 
     @Test
-    fun `should support WITHIN expression with named parameter as collection`() {
+    fun `should support NOT WITHIN expression with named parameter as collection`() {
         val parameterValue = listOf(1, 2, 3)
         val parameterName = "param"
         val expected = DopeQuery(
-            queryString = "`numberField` WITHIN \$$parameterName",
+            queryString = "`numberField` NOT WITHIN \$$parameterName",
             DopeParameters(namedParameters = mapOf(parameterName to parameterValue)),
         )
-        val underTest = WithinExpression(someNumberField(), parameterValue.asParameter(parameterName))
+        val underTest = NotWithinExpression(someNumberField(), parameterValue.asParameter(parameterName))
 
         val actual = underTest.toDopeQuery(manager)
 
@@ -81,13 +81,13 @@ class WithinExpressionTest : ManagerDependentTest {
     }
 
     @Test
-    fun `should support WITHIN expression with positional parameter as collection`() {
+    fun `should support NOT WITHIN expression with positional parameter as collection`() {
         val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            queryString = "`numberField` WITHIN $1",
+            queryString = "`numberField` NOT WITHIN $1",
             DopeParameters(positionalParameters = listOf(parameterValue)),
         )
-        val underTest = WithinExpression(someNumberField(), parameterValue.asParameter())
+        val underTest = NotWithinExpression(someNumberField(), parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery(manager)
 
@@ -95,16 +95,16 @@ class WithinExpressionTest : ManagerDependentTest {
     }
 
     @Test
-    fun `should support WITHIN expression with named parameters as value and collection`() {
+    fun `should support NOT WITHIN expression with named parameters as value and collection`() {
         val parameterValue = 1
         val parameterCollectionValue = listOf(1, 2, 3)
         val parameterNameA = "paramA"
         val parameterNameB = "paramB"
         val expected = DopeQuery(
-            queryString = "\$$parameterNameA WITHIN \$$parameterNameB",
+            queryString = "\$$parameterNameA NOT WITHIN \$$parameterNameB",
             DopeParameters(namedParameters = mapOf(parameterNameA to parameterValue, parameterNameB to parameterCollectionValue)),
         )
-        val underTest = WithinExpression(parameterValue.asParameter(parameterNameA), parameterCollectionValue.asParameter(parameterNameB))
+        val underTest = NotWithinExpression(parameterValue.asParameter(parameterNameA), parameterCollectionValue.asParameter(parameterNameB))
 
         val actual = underTest.toDopeQuery(manager)
 
@@ -112,14 +112,14 @@ class WithinExpressionTest : ManagerDependentTest {
     }
 
     @Test
-    fun `should support WITHIN expression with positional parameters as value and collection`() {
+    fun `should support NOT WITHIN expression with positional parameters as value and collection`() {
         val parameterValue = 1
         val parameterCollectionValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            queryString = "$1 WITHIN $2",
+            queryString = "$1 NOT WITHIN $2",
             DopeParameters(positionalParameters = listOf(parameterValue, parameterCollectionValue)),
         )
-        val underTest = WithinExpression(parameterValue.asParameter(), parameterCollectionValue.asParameter())
+        val underTest = NotWithinExpression(parameterValue.asParameter(), parameterCollectionValue.asParameter())
 
         val actual = underTest.toDopeQuery(manager)
 
@@ -127,89 +127,89 @@ class WithinExpressionTest : ManagerDependentTest {
     }
 
     @Test
-    fun `should support WITHIN extension type type`() {
+    fun `should support NOT WITHIN extension type type`() {
         val value = someNumberField()
         val collection = someNumberArrayField()
-        val expected = WithinExpression(value, collection)
+        val expected = NotWithinExpression(value, collection)
 
-        val actual = value.withinArray(collection)
+        val actual = value.notWithinArray(collection)
 
         assertEquals(expected.toDopeQuery(manager), actual.toDopeQuery(manager))
     }
 
     @Test
-    fun `should support WITHIN extension number type`() {
+    fun `should support NOT WITHIN extension number type`() {
         val value = 1
         val collection = someNumberArrayField()
-        val expected = WithinExpression(value.toDopeType(), collection)
+        val expected = NotWithinExpression(value.toDopeType(), collection)
 
-        val actual = value.withinArray(collection)
+        val actual = value.notWithinArray(collection)
 
         assertEquals(expected.toDopeQuery(manager), actual.toDopeQuery(manager))
     }
 
     @Test
-    fun `should support WITHIN extension string type`() {
+    fun `should support NOT WITHIN extension string type`() {
         val value = "s"
         val collection = someStringArrayField()
-        val expected = WithinExpression(value.toDopeType(), collection)
+        val expected = NotWithinExpression(value.toDopeType(), collection)
 
-        val actual = value.withinArray(collection)
+        val actual = value.notWithinArray(collection)
 
         assertEquals(expected.toDopeQuery(manager), actual.toDopeQuery(manager))
     }
 
     @Test
-    fun `should support WITHIN extension boolean type`() {
+    fun `should support NOT WITHIN extension boolean type`() {
         val value = true
         val collection = someBooleanArrayField()
-        val expected = WithinExpression(value.toDopeType(), collection)
+        val expected = NotWithinExpression(value.toDopeType(), collection)
 
-        val actual = value.withinArray(collection)
+        val actual = value.notWithinArray(collection)
 
         assertEquals(expected.toDopeQuery(manager), actual.toDopeQuery(manager))
     }
 
     @Test
-    fun `should support WITHIN extension type type collection`() {
+    fun `should support NOT WITHIN extension type type collection`() {
         val value = someNumberField()
         val collection = listOf(someNumberField(), someNumberField())
-        val expected = WithinExpression(value, collection.toDopeType())
+        val expected = NotWithinExpression(value, collection.toDopeType())
 
-        val actual = value.withinArray(collection)
+        val actual = value.notWithinArray(collection)
 
         assertEquals(expected.toDopeQuery(manager), actual.toDopeQuery(manager))
     }
 
     @Test
-    fun `should support WITHIN extension number type collection`() {
+    fun `should support NOT WITHIN extension number type collection`() {
         val value = 1
         val collection = listOf(someNumberField(), someNumberField())
-        val expected = WithinExpression(value.toDopeType(), collection.toDopeType())
+        val expected = NotWithinExpression(value.toDopeType(), collection.toDopeType())
 
-        val actual = value.withinArray(collection)
+        val actual = value.notWithinArray(collection)
 
         assertEquals(expected.toDopeQuery(manager), actual.toDopeQuery(manager))
     }
 
     @Test
-    fun `should support WITHIN extension string type collection`() {
+    fun `should support NOT WITHIN extension string type collection`() {
         val value = "s"
         val collection = listOf(someStringField(), someStringField())
-        val expected = WithinExpression(value.toDopeType(), collection.toDopeType())
+        val expected = NotWithinExpression(value.toDopeType(), collection.toDopeType())
 
-        val actual = value.withinArray(collection)
+        val actual = value.notWithinArray(collection)
 
         assertEquals(expected.toDopeQuery(manager), actual.toDopeQuery(manager))
     }
 
     @Test
-    fun `should support WITHIN extension boolean type collection`() {
+    fun `should support NOT WITHIN extension boolean type collection`() {
         val value = true
         val collection = listOf(someBooleanField(), someBooleanField())
-        val expected = WithinExpression(value.toDopeType(), collection.toDopeType())
+        val expected = NotWithinExpression(value.toDopeType(), collection.toDopeType())
 
-        val actual = value.withinArray(collection)
+        val actual = value.notWithinArray(collection)
 
         assertEquals(expected.toDopeQuery(manager), actual.toDopeQuery(manager))
     }
@@ -218,9 +218,9 @@ class WithinExpressionTest : ManagerDependentTest {
     fun `should support WITHIN extension type select`() {
         val value = someStringField()
         val selectClause = someSelectRawClause()
-        val expected = WithinExpression(value, selectClause.asExpression())
+        val expected = NotWithinExpression(value, selectClause.asExpression())
 
-        val actual = value.withinArray(selectClause)
+        val actual = value.notWithinArray(selectClause)
 
         assertEquals(expected.toDopeQuery(manager), actual.toDopeQuery(manager))
     }
@@ -229,9 +229,9 @@ class WithinExpressionTest : ManagerDependentTest {
     fun `should support WITHIN extension number select`() {
         val value = 1
         val selectClause = someNumberSelectRawClause()
-        val expected = WithinExpression(value.toDopeType(), selectClause.asExpression())
+        val expected = NotWithinExpression(value.toDopeType(), selectClause.asExpression())
 
-        val actual = value.withinArray(selectClause)
+        val actual = value.notWithinArray(selectClause)
 
         assertEquals(expected.toDopeQuery(manager), actual.toDopeQuery(manager))
     }
@@ -240,9 +240,9 @@ class WithinExpressionTest : ManagerDependentTest {
     fun `should support WITHIN extension string select`() {
         val value = "s"
         val selectClause = someStringSelectRawClause()
-        val expected = WithinExpression(value.toDopeType(), selectClause.asExpression())
+        val expected = NotWithinExpression(value.toDopeType(), selectClause.asExpression())
 
-        val actual = value.withinArray(selectClause)
+        val actual = value.notWithinArray(selectClause)
 
         assertEquals(expected.toDopeQuery(manager), actual.toDopeQuery(manager))
     }
@@ -251,9 +251,9 @@ class WithinExpressionTest : ManagerDependentTest {
     fun `should support WITHIN extension boolean select`() {
         val value = true
         val selectClause = someBooleanSelectRawClause()
-        val expected = WithinExpression(value.toDopeType(), selectClause.asExpression())
+        val expected = NotWithinExpression(value.toDopeType(), selectClause.asExpression())
 
-        val actual = value.withinArray(selectClause)
+        val actual = value.notWithinArray(selectClause)
 
         assertEquals(expected.toDopeQuery(manager), actual.toDopeQuery(manager))
     }
