+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/collection/SatisfiesTest.kt
@@ -1,4 +1,4 @@
-package ch.ergon.dope.operators.collection
+package ch.ergon.dope.resolvable.expression.unaliased.type.collection
 
 import ch.ergon.dope.DopeQuery
 import ch.ergon.dope.DopeQueryManager
